requirements.txt
virtualenv
flask
flask_sqlalchemy
mysql-connector-python
python-dotenv